+++ src/main/java/io/github/jsousa32/libdealsign/usecases/signatures/models/common/Status.java
@@ -1,4 +1,4 @@
-package io.github.jsousa32.libdealsign.usecases.signatures.models.create;
+package io.github.jsousa32.libdealsign.usecases.signatures.models.common;
 
 public enum Status {
     LINKED,
